+++ NEWS.md
@@ -1,4 +1,4 @@
-# asciicast (development version)
+# asciicast 2.3.0
 
 * `record()` has a new argument: `show_output`, to show the output from the
   subprocess in real time.